Fix id with non imprimible characters
Because the id of the resources were computed using a sha1 hash of the
contents of the .env file, it could contain non imprimible characters.

This commit attempts to fix the problem described above. Instead of
using directly the sha1 hash, we can encode it in base64 and be sure it
only contains safe imprimible characters.